wddm: autoresize fixes
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@34130 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/include/VBoxDisplay.h
@@ -32,7 +32,7 @@
 
 typedef struct
 {
-    ULONG Id;
+    DWORD Id;
     DWORD Width;
     DWORD Height;
     DWORD BitsPerPixel;
